+++ src/index.ts
@@ -3,4 +3,4 @@ export type { DeepPartial, OperationMock, Patch } from "./types.ts";
 
 export { codegen } from "./codegen.ts";
 export { clear, operation, proxy } from "./proxy.ts";
-export { formatCode, toObject } from "./util.ts";
+export { toObject } from "./util.ts";