+++ src/jest.polyfill.test.ts
@@ -1,31 +1,14 @@
-import { afterEach, beforeEach } from "jsr:@std/testing/bdd";
+import { afterEach } from "jsr:@std/testing/bdd";
 
 // deno-lint-ignore no-explicit-any
 globalThis.afterEach = afterEach as any;
 
-declare namespace jasmine {
-  type CustomReporterResult = {
-    failedExpectations: unknown[];
-  };
-
-  const getEnv: () => {
-    addReporter: (
-      props: { specStarted: (result: jasmine.CustomReporterResult) => void },
-    ) => void;
-  };
-}
-
-const reporters: ((result: unknown) => void)[] = [];
-beforeEach(() => {
-  const result = { failedExpectations: [] };
-  for (const reporter of reporters) reporter(result);
-});
-globalThis.jasmine = {
-  // @ts-ignore Declared it
-  getEnv: () => ({
-    addReporter: (
-      { specStarted }: { specStarted: (result: unknown) => void },
-    ) => reporters.push(specStarted),
+// deno-lint-ignore no-explicit-any
+(globalThis as any).expect = {
+  getState: () => ({
+    assertionCalls: 0,
+    numPassingAsserts: 0,
+    suppressedErrors: [],
   }),
 };
 
@@ -44,29 +27,25 @@ if (!globalThis.jest.fn) {
   }) as typeof jest["fn"];
 }
 
-const path = `${
-  Deno.env.get("HOME")
-}/Library/Caches/deno/npm/registry.npmjs.org/graphql-tag/2.12.6/package.json`;
-const packageJson = JSON.parse(await Deno.readTextFile(path));
+const graphqlTagPaths = [
+  `${
+    Deno.env.get("HOME")
+  }/.cache/deno/npm/registry.npmjs.org/graphql-tag/2.12.6/package.json`,
+  `${
+    Deno.env.get("HOME")
+  }/Library/Caches/deno/npm/registry.npmjs.org/graphql-tag/2.12.6/package.json`,
+];
 
-if (packageJson.main !== "./lib/index.js") {
-  packageJson.main = "./lib/index.js"; // Fix the main entry
+for (const path of graphqlTagPaths) {
+  try {
+    const packageJson = JSON.parse(await Deno.readTextFile(path));
 
-  await Deno.writeTextFile(path, JSON.stringify(packageJson, null, 2));
+    if (packageJson.main !== "./lib/index.js") {
+      packageJson.main = "./lib/index.js"; // Fix the main entry
 
-  console.log("Patched graphql-tag package.json successfully.");
+      await Deno.writeTextFile(path, JSON.stringify(packageJson, null, 2));
+
+      console.log("Patched graphql-tag package.json successfully.");
+    }
+  } catch {}
 }
-
-let _fail: unknown;
-globalThis.fail = (error) => {
-  _fail = error instanceof Error ? error : Object.assign(new Error(), error);
-  throw error;
-};
-
-afterEach(() => {
-  if (_fail) {
-    const error = _fail;
-    _fail = undefined;
-    throw error;
-  }
-});